feat: 사이트 공통 에러 핸들링, 중앙 store zustand 작업

+++ app/components/survey/SurveyListItem.tsx
@@ -1,30 +1,37 @@
+import { Link } from 'react-router'
 import Button from '~/components/common/Button'
 import Title from '~/components/common/Title'
-import { deleteSurvey } from '~/lib/surveyStorage'
+import ALERTS from '~/constants/alerts'
 import { cn } from '~/lib/utils'
+import useErrorStore from '~/store/errorStore'
+import useSurveyListStore from '~/store/surveyListStore'
 import type { SurveyData } from '~/types/survey'
 
 type SurveyListItemProps = {
 	survey: SurveyData
-	setSurveyList: React.Dispatch<React.SetStateAction<SurveyData[] | null>>
 }
 
 const baseShapeStyle = 'rounded-md h-[80px]'
 
-export default function SurveyListItem({
-	survey,
-	setSurveyList
-}: SurveyListItemProps) {
+export default function SurveyListItem({ survey }: SurveyListItemProps) {
+	const { deleteSurvey } = useSurveyListStore()
+	const { handleError } = useErrorStore()
+
 	const onClickDeleteForm = (surveyId: SurveyData['id']) => {
-		const newSurveyList = deleteSurvey(surveyId)
-		if (!newSurveyList) return
-		setSurveyList(newSurveyList)
+		const isConfirmed = confirm(ALERTS.CONFIRM.DELETE_FORM)
+		if (!isConfirmed) return
+		try {
+			deleteSurvey(surveyId)
+			alert(ALERTS.SUCCESS.DELETE_FORM)
+		} catch (e) {
+			handleError(e)
+		}
 	}
 
 	return (
 		<li className={cn('border-primary relative border', baseShapeStyle)}>
-			<a
-				href={`/${survey.id}/edit`}
+			<Link
+				to={`/${survey.id}/edit`}
 				className="bg-primary/10 hover:bg-primary/30 block w-full p-4"
 			>
 				<Title h="h3" size="sm">
@@ -33,7 +40,7 @@ export default function SurveyListItem({
 				<time dateTime={survey.createdAt} className="text-xs text-gray-300">
 					{new Date(survey.createdAt).toLocaleString('ko-KR')}
 				</time>
-			</a>
+			</Link>
 			<div className="absolute right-0 bottom-0 flex justify-end">
 				<Button
 					size="sm"
